+++ crates/scheduler/src/call/crypto/ccm.rs
@@ -13,36 +13,36 @@
 // limitations under the License.
 
 use wasefire_applet_api::crypto::ccm::{self as api, Api};
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 use wasefire_board_api::crypto::aead::{Api as _, Array};
 use wasefire_board_api::Api as Board;
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 use wasefire_board_api::{self as board, Support};
 
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 use crate::applet::store::MemoryApi;
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 use crate::Trap;
 use crate::{DispatchSchedulerCall, SchedulerCall};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::IsSupported(call) => is_supported(call),
-        Api::Encrypt(call) => or_trap!("board-api-aes128-ccm", encrypt(call)),
-        Api::Decrypt(call) => or_trap!("board-api-aes128-ccm", decrypt(call)),
+        Api::Encrypt(call) => or_trap!("board-api-crypto-aes128-ccm", encrypt(call)),
+        Api::Decrypt(call) => or_trap!("board-api-crypto-aes128-ccm", decrypt(call)),
     }
 }
 
 fn is_supported<B: Board>(call: SchedulerCall<B, api::is_supported::Sig>) {
     let api::is_supported::Params {} = call.read();
-    #[cfg(feature = "board-api-aes128-ccm")]
+    #[cfg(feature = "board-api-crypto-aes128-ccm")]
     let supported = bool::from(board::crypto::Aes128Ccm::<B>::SUPPORT) as u32;
-    #[cfg(not(feature = "board-api-aes128-ccm"))]
+    #[cfg(not(feature = "board-api-crypto-aes128-ccm"))]
     let supported = 0;
     call.reply(Ok(api::is_supported::Results { supported: supported.into() }))
 }
 
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 fn encrypt<B: Board>(mut call: SchedulerCall<B, api::encrypt::Sig>) {
     let api::encrypt::Params { key, iv, len, clear, cipher } = call.read();
     let scheduler = call.scheduler();
@@ -61,7 +61,7 @@ fn encrypt<B: Board>(mut call: SchedulerCall<B, api::encrypt::Sig>) {
     call.reply(results);
 }
 
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 fn decrypt<B: Board>(mut call: SchedulerCall<B, api::decrypt::Sig>) {
     let api::decrypt::Params { key, iv, len, cipher, clear } = call.read();
     let scheduler = call.scheduler();
@@ -81,12 +81,12 @@ fn decrypt<B: Board>(mut call: SchedulerCall<B, api::decrypt::Sig>) {
     call.reply(results);
 }
 
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 fn expand_iv(iv: &[u8]) -> Array<typenum::U13> {
     core::array::from_fn(|i| i.checked_sub(5).map(|i| iv[i]).unwrap_or(0)).into()
 }
 
-#[cfg(feature = "board-api-aes128-ccm")]
+#[cfg(feature = "board-api-crypto-aes128-ccm")]
 fn ensure_support<B: Board>() -> Result<(), Trap> {
     match bool::from(board::crypto::Aes128Ccm::<B>::SUPPORT) {
         true => Ok(()),